+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2025-05-17
+
+### Added
+
+- Added support for "`cn.unity.physical-light-unit`" (external URP package).
+
+### Fixed
+
+- Fixed a rendering issue with dynamic resolution enabled.
+
+
 ## [1.0.3] - 2025-04-02
 
 ### Fixed